refactor(architecture): move lint baseline into extension domain
